requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
cffi==1.10.0
constantly==15.1.0
cryptography==1.9
cssselect==1.0.1
Django==1.11.2
djangorestframework==3.6.3
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
lxml==3.8.0
mysqlclient==1.3.10
parsel==1.2.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==17.0.0
pytz==2017.2
queuelib==1.4.2
Scrapy==1.4.0
service-identity==17.0.0
six==1.10.0
Twisted==17.5.0
w3lib==1.17.0
zope.interface==4.4.2
